+++ contracts/orderbook/OrderBookStateBuilder.sol
@@ -15,7 +15,7 @@ contract OrderBookStateBuilder is StandardStateBuilder {
         Operand,
         StackTop stackTop_
     ) internal pure returns (StackTop) {
-        function(uint) internal view returns (uint) fn_;
+        function(uint256) internal view returns (uint256) fn_;
         return integrityState_.applyFn(stackTop_, fn_);
     }
 
@@ -24,7 +24,7 @@ contract OrderBookStateBuilder is StandardStateBuilder {
         Operand,
         StackTop stackTop_
     ) internal pure returns (StackTop) {
-        function(uint, uint) internal view returns (uint) fn_;
+        function(uint256, uint256) internal view returns (uint256) fn_;
         return integrityState_.applyFn(stackTop_, fn_);
     }
